Adds license defaulting to ASL2
 * By default, the project set the ASL2 license
 * Adds a new model for specifying other licenses
 * Adds new scripted test cases
