+++ .gitignore
@@ -5,3 +5,4 @@ target/
 deploy_key
 .DS_Store
 output/*
+/secring.gpg
