+++ src/sbt-test/sbtorgpolicies/custom-template-config/build.sbt
@@ -1,4 +1,4 @@
-import sbtorgpolicies._
+import sbtorgpolicies.model._
 
 scalaVersion := "2.11.8"
 
@@ -8,7 +8,7 @@ orgGithubSettings := GitHubSettings(
   organizationName = "My Organization",
   organizationHomePage = url("http://myorg.com"),
   organizationEmail = "hello@myorg.com",
-  license = apache
+  license = ApacheLicense.custom
 )
 
 orgTemplatesDirectory := (resourceDirectory in Compile).value / "myorg"
@@ -18,11 +18,35 @@ def getLines(fileName: String) =
 
 lazy val check = TaskKey[Unit]("check")
 
+lazy val checkApache = TaskKey[Unit]("checkApache")
+
+lazy val checkMIT = TaskKey[Unit]("checkMIT")
+
 check := {
   val content = getLines("LICENSE").mkString
 
   if (!content.startsWith("Custom License"))
     sys.error("user template is not being copied successfully")
+
+  if (!content.contains("My Organization"))
+    sys.error("custom user setting didn't work")
+}
+
+checkApache := {
+  val content = getLines("LICENSE").mkString
+
+  if (!content.startsWith("Copyright ") || !content.contains("Licensed under the Apache License, Version 2.0"))
+    sys.error("Apache template is not being copied successfully")
+
+  if (!content.contains("My Organization"))
+    sys.error("custom user setting didn't work")
+}
+
+checkMIT := {
+  val content = getLines("LICENSE").mkString
+
+  if (!content.startsWith("The MIT License (MIT)"))
+    sys.error("MIT template is not being copied successfully")
 
   if (!content.contains("My Organization"))
     sys.error("custom user setting didn't work")
