+++ src/sbt-test/sbtorgpolicies/custom-template/src/main/resources/org/templates/LICENSE_ASL2.template
